DATA-575 Add custom Transformers, PipelineModel, PipelineEnsemble
